### 2019-10-04 #1782 @kyleniemeyer
Hi @mjsottile, how do you feel about taking on this submission as your first editorial assignment?
