### 2019-10-04 eaecca2
Update paper.md
Fix references style
