+++ joss/paper.bib
@@ -18,15 +18,6 @@
     url = {http://papers.nips.cc/paper/7892-neural-ordinary-differential-equations.pdf}
 }
 
-@misc{ANODE:2019,
-    title={Augmented Neural ODEs},
-    author={Emilien Dupont and Arnaud Doucet and Yee Whye Teh},
-    year={2019},
-    eprint={1904.01681},
-    archivePrefix={arXiv},
-    primaryClass={stat.ML}
-}
-
 @article{VBS:2019,
     author={Mats Vermeeren and Alessandro Bravetti and Marcello Seri},
     title={Contact variational integrators},
