### 2019-10-08 #1782 @mjsottile
Some potential reviewers are @rljacobson, @xoolive, @ziotom78, and @ocramz.  Would any of you be interested in reviewing this paper?
