### 2019-10-16 #1782 @dagit
@mjsottile I'd be happy to take a look at this. I'm not an Ocaml expert but I have experience in other functional languages.
