### 2019-10-16 #1812 @whedon
[REVIEW]: OwlDE: making ODEs first-class `Owl` citizens
**Submitting author:** @mseri (<a href="http://orcid.org/0000-0002-8563-5894">Marcello Seri</a>)
**Repository:** <a href="https://github.com/owlbarn/owl_ode/" target ="_blank">https://github.com/owlbarn/owl_ode/</a>
**Version:** v0.2.0
