### 2019-10-25 #1812 @xoolive
Hello @mseri

I have first remarks about the current state of the repository/project.

- Why is the name of the paper OwlDE when the whole project name + repository + documentation says owl-ode (which is much clearer to me to understand what it is about)
- `opam install owl-ode` does not work here, even after `opam update`. Should we set a specific repository?
- There are many rendering issues (markdown and LaTeX rendering) on the documentation pages, both ocaml.xyz/apidoc/ and ocaml.xyz/owl_ode
