### 2019-10-25 #1812 @mseri
Hi @xoolive. Let me reply what I can:

- We changed the name in view of the official presentation of the library during ICFP 2019. The name OwlDE was chosen because it was a contraction of Owl and ODE. The underlying library name was not changed to avoid backward incompatibilities for the moment. The current plan was to rename it to owlde with a tansition owl-ode package once owl reaches 1.0 (which is currently not a known deadline).

- This is surprising. Unless it is owl failing (there is a known problem of ubuntu's openblas that requires manual compilation of it), owl-ode should compile without problems. It is tested on multiple distributions, on ubuntu with custom openblas, and on macosx before each release, and tested by the ocaml-ci each time is released on opam. If you can provide me the logs, I can see what is wrong

- Thanks for pointing out the issues. It is a bug in the treatment of latex in the ocaml documentation generation for which we have a workaround (I forgot to apply it during the latest documentation release). Will be fixed by monday. The apidoc issue is known and is a problem of how owl deals with documentation. This is being redesigned currently, and will be automatically fixed once owl moves to the new documentation structure.

I
