### 2019-10-25 #1812 @xoolive
@mseri In case it helps

```
$ opam update
[...]
$ opam install owl-ode
[ERROR] No package named owl-ode found.
[1]    17437 exit 66    opam install owl-ode
```

May I suggest you check:
```
opam repository list
```

Mine is:
```
  10 [http]  coq-released     https://coq.github.io/opam-coq-archive/released-opam1.2
   0 [http]     default     https://opam.ocaml.org/1.2.2
```
