### 2019-10-25 #1812 @mseri
Oh I see the problem. Opam has moved to opam2 since almost one year. For some reason you are using the old unmaintained repository (and potentially still opam 1).  Coq is released in the new opam2 repository as well, you should consider upgrading the system.
