### 2019-10-25 #1812 @mseri
You can find the information about the move to opam2 and how to upgrade here: https://opam.ocaml.org/doc/Upgrade_guide.html
