### 2019-11-03 #1812 @xoolive
Sorry for the delay in coming back to this. I upgraded opam and tried again.

It seems the installation work but not the examples nor the tests due to some linking issues.

I would advocate for more detailed installation instructions and better dependency management. You will find below all I tried before giving up.

Sorry, I am not willing to dig more alone, by myself, into these linking issues. No problem to try more according to your instructions.

I think a troubleshooting section including opam and dependencies appearing on the github or in the documentation will help me tick the box 😉

```
opam switch create 4.06.0
opam install owl-ode  # failed
sudo apt install liblapacke-dev libopenblas-dev
# try again => ok
```
```
dune exec examples/van_der_pol.exe  # fail, file does not exist
```
```
dune exec examples/van_der_pol_odepack.exe
opam install owl-plplot  # fail
sudo apt install libplplot-dev libshp-dev
# opam ok
# dune dependency failed again
opam install odepack
sudo apt install gfortran
# opam ok
```
```
# dune failing with linking issue (traceback)

/home/xo/.opam/4.06.0/.opam-switch/build/owl.0.6.0/_build/default/src/owl/src/owl/lapacke/owl_lapacke_generated_stub.c:33773: undefined reference to `LAPACKE_zlagsy'
collect2: error: ld returned 1 exit status
File "caml_startup", line 1:
Error: Error during linking
[1]    29805 exit 1     dune exec examples/van_der_pol_odepack.exe
```
```
dune exec examples/damped.exe  # fail, file with same linking issue
```
```
dune exec tests/tests.exe  # fail with dependency issue
opam install alcotest
# try again and fail with same linking issue
```
