### 2019-11-04 #1812 @mseri
Dear @xoolive thanks for the update. The linking issue is a problem of owl and how (usually `ubuntu` derivatives) package `openblas`. I have added a more explicit troubleshooting section on `owl` readme and will make a PR for `owl`'s documentation website about it in the next few days. I have also updated the README of `owl_ode`'s repository to [link to it and be more explicit about the dependencies](https://github.com/owlbarn/owl_ode/blob/master/README.md#owlde---ordinary-differential-equation-solvers-).

To make sure you installed all the dependencies, in most cases, you can make `opam` deal with it for you with `opam depext owl-ode owl-ode-odepack owl-ode-sundials` (in fact `opam depext owl-ode-sundials` should be enough). This is not enough in ubuntu's case, but there is no will to change it at ubuntu level, and we are not yet shipping our custom `openblas` in `owl` (as done by e.g. numpy).

The sundials bindings use sundials3, so sundials may have to be compiled manually (the low-level bindings are provided by https://github.com/inria-parkas/sundialsml, the documentation there provides the install and troubleshooting instructions and is now linked in owl-ode readme).

To install (and run the tests) you need to ask opam for the tests dependencies: `opam install -t owl-ode owl-ode-odepack owl-ode-sundials`. This will also run the unit test before installing the package.

Thanks for looking into it, I was not expecting for you to dig through these issues alone. Until now all of our users were already familiar with either the general `ocaml/opam` `dune`-based workflow or `owl` itself: your feedback as an external user is invaluable.
