### 2019-11-06 #1812 @xoolive
OK I managed to recompile openblas and run the examples. I have ticked all the boxes, which thereby finishes the reviewing process. I wish you the best in this project. Thank you for taking the comments into account. 👍

In general, I would recommend to document the installation process more thoroughly (you addressed my comments, but I would definitely not call the result a "smooth installation process"). Maybe this comment does not particularly apply to owl-ode but more to the whole owl project. You can be helpful to a handful of users, but how long are you willing to make it?

For a newcomer to the project who could say "why not switch to OCaml because this project is nice?", there would be a lot of reasons to give up (and I didn't try to install on Windows 😆)
