### 2019-11-07 #1812 @mseri
@xoolive thank you very much for the review, the feedback and the patience. I fully agree with your comment: Installation and distribution of owl are one of the main pain points at the moment. We are currently working on revamping its documentation system and the tutorial book, also to address the installation procedure.

The current one, available here https://ocaml.xyz/chapter/index.html, is getting out of date and is hard to maintain. The new one ( in development here https://github.com/owlbarn/owl_tutorials and on the personal forks ) will run all the code snippets as unit test and will run on docker images that are built with the instructions from the new install section. Hopefully this will help further in the smoothening of the install procedure and the onboarding on new users.

We have also considered bundling openblas itself and statically link it, but it makes the compilation time much longer. Unfortunately at the moment it is not possible to distribute binary artifacts in Ocaml. There is ongoing work to change this at the level of the compiler but it’s not going to happen soon as far as I know.
