### 2019-11-07 #1812 @mjsottile
@xoolive - thanks for the review!

@mseri - I agree regarding distributing OCaml packages with external dependencies.  It can be challenging but your package appears to be similar to others in terms of difficulty of installation.  As for Windows, the use of WSL seems to be becoming more popular, so I typically recommend people go that route instead of attempting to run OCaml natively.

@rljacobson and @dagit , let me know if you need any assistance in completing the review.
