### 2019-12-06 #1812 @mjsottile
I removed the reviewer who went completely unresponsive.  I left @rljacobson for now as they have at least checked a few boxes, but appear to have stopped responding.

I am reaching out to a couple other potential reviewers off of GitHub.  @mseri - do you have any suggested people I should try to recruit?  If I cannot find an additional reviewer, I will do the review myself when I return from travel so this submission doesn't linger longer than it needs to.

Apologies for the delays with getting reviews completed so far.
