### 2019-12-12 5579263
add .dev to pinning in travis to avoid version confusion
Signed-off-by: Marcello Seri <marcello.seri@gmail.com>
