+++ .travis.yml
@@ -6,7 +6,7 @@ install: wget https://raw.githubusercontent.com/ocaml/ocaml-travisci-skeleton/ma
 script: bash -ex ./.travis-docker.sh
 env:
   global:
-  - PINS="owl-ode-base:. owl-ode.0.1.0:. owl-ode-odepack:. owl-ode-sundials:."
+  - PINS="owl-ode-base.dev:. owl-ode.dev:. owl-ode-odepack.dev:. owl-ode-sundials.dev:."
   matrix:
   - PACKAGE="owl-ode-base" DISTRO="debian-stable" OCAML_VERSION="4.07"
   - PACKAGE="owl-ode" DISTRO="debian-stable" OCAML_VERSION="4.07"
