### 2019-12-13 #1812 @mjsottile
@mseri I have added myself as a reviewer and am working through the review process.  In the meantime, I have issued a pull request to address a typo in the paper.  Also, can you add DOIs for the references in the paper?  It does not appear that any of the referenced papers have DOIs.

@xoolive Can you confirm that you are satisfied with the review and would recommend acceptance?  Sorry for the long delay since you finished your review.
