### 2019-12-15 cd0ef42
Add doi where possible
Signed-off-by: Marcello Seri <marcello.seri@gmail.com>
