+++ joss/paper.bib
@@ -2,9 +2,7 @@
     title={Owl: A General-Purpose Numerical Library in OCaml},
     author={Liang Wang},
     year={2017},
-    eprint={1707.09616},
-    archivePrefix={arXiv},
-    primaryClass={cs.MS}
+    eprint = {arXiv:1707.09616}
 }
 
 @incollection{NODE:2018,
@@ -23,5 +21,6 @@
     title={Contact variational integrators},
     journal={Journal of Physics A: Mathematical and Theoretical},
     url={http://iopscience.iop.org/10.1088/1751-8121/ab4767},
-    year={2019}
+    year={2019},
+    doi={10.1088/1751-8121/ab4767}
 }
