### 2019-12-15 4a0962a
Add url to Liang whitepaper
Signed-off-by: Marcello Seri <marcello.seri@gmail.com>
