+++ joss/paper.bib
@@ -2,7 +2,10 @@
     title={Owl: A General-Purpose Numerical Library in OCaml},
     author={Liang Wang},
     year={2017},
-    eprint = {arXiv:1707.09616}
+    eprint={1707.09616},
+    archivePrefix={arXiv},
+    primaryClass={cs.MS},
+    url = {https://arxiv.org/abs/1707.09616}
 }
 
 @incollection{NODE:2018,
