### 2019-12-15 #1812 @mseri
I have added the only available DOI:

- I could not find a DOI for the Neural ODE paper, the bibtex entry is the official one from https://papers.nips.cc/paper/7892-neural-ordinary-differential-equations

- The Owl whitepaper does not have a DOI. It was presented at ICFP 2017, https://icfp17.sigplan.org/details/ocaml-2017-talks/5/Owl-A-General-Purpose-Numerical-Library-in-OCaml, I have used the bibtex entry generated by arxiv.
