### 2019-12-19 #1812 @mjsottile
I've completed reviewing the project - looks good to me.  It installed fine and I was able to run the tests and work through the examples and documentation.  Given that @xoolive indicated they were done with their review and satisfied, we have two reviews now with mine.

The final step in the process before I hand it off to the EIC team is to make a tagged release and archive.  JOSS needs the software to be archived in an archival repository, such as Zenodo, figshare, or an institutional repository. Once this is done, that repository should provide a DOI to the archive.  Please post the DOI that you receive in this thread so I can add it to the submission.  For software in GitHub, the easiest path is usually to use to the Zenodo-GitHub linkage, being sure to manually change the metadata in Zenodo so that the title and authors of the archive match the those of the JOSS paper.

Once you have completed this, I'll run the final steps and hand the submission off to get it finished up.
