### 2019-12-19 779e0b7
Add JOSS paper to latest released version
Signed-off-by: Marcello Seri <marcello.seri@gmail.com>
