+++ joss/paper.md
@@ -1,0 +1,102 @@
+---
+title: 'OwlDE: making ODEs first-class `Owl` citizens'
+tags:
+    - OCaml
+    - Ordinary differential equations
+    - Geometric integration
+    - Numerical integration
+    - Neural ODE
+authors:
+    - name: Marcello Seri
+      orcid: 0000-0002-8563-5894
+      affiliation: 1
+    - name: Ta-Chu Kao
+      affiliation: 2
+affiliations:
+    - name:  Bernoulli Institute for Mathematics, Computer Science and Artificial Intelligence, University of Groningen
+      index: 1
+    - name: Computational and Biological Learning Lab, Department of Engineering, University of Cambridge
+      index: 2
+date: 4 October 2019
+bibliography: paper.bib
+---
+
+# Summary
+
+After only three years of intensive development and continuous optimisation, `Owl` has emerged in the `OCaml` ecosystem as a versatile and powerful scientific programming library, competitive with mainstream libraries such as `SciPy` and `NumPy`.
+What sets `Owl` apart is that it brings under one umbrella the flexibility of a dynamical language and the power and safety of the `OCaml` type system [@Wang:2017].
+
+Today, `Owl` can be used to solve a wide range of scientific problems: it provides efficient types for handling multidimensional arrays and linear algebra operations built on top of `BLAS` and `LAPACK`; it supports machine learning applications with a powerful computational graph engine and automatic differentiation pipeline.
+To improve efficiency, `Owl` allows offloading computations to distributed systems and GPUs.
+With the recent addition of `dataframes` and integration with Jupyter Notebooks provided by `ocaml-jupyter`, `Owl` has the chance to become an excellent framework for exploratory mathematical analysis.
+
+A notable omission in `Owl`'s ecosystem, when compared to similar solutions in `python` and `Julia`, was a package for solving ordinary differential equations.
+To fill this need, we designed `OwlDE`, a flexible and efficient ODE engine for `Owl`.
+
+## Design of the core module
+
+The lack of automation around type classes and dynamic typing in `OCaml` may seem at first a huge impediment to designing a flexible and ergonomic ODE integrator library.
+Indeed, such a library should ideally allow the users to seamlessly use different algorithms that require different kinds of inputs and options, and return varying kinds of outputs.
+Such constraints on input/output types pose a major problem for the strong static type system of `OCaml`, even if they are in principle implementable exploiting advanced language features, and were one of the central issues in designing `OwlDE`.
+
+We iterated various options and settled on the use of first-class modules.
+These have been introduced in OCaml since version 3.12, and further improved in subsequent releases.
+With first-class modules the user can parametrise functions over modules, allowing us to find a middle ground between the verbosity of functorial code and the composability of OCaml functions.
+The flexibility of this API allowed us to provide integration and bindings to external frameworks like `SUNDIALS` or `ODEPACK` in a completely seamless way.
+Such an API allowed us to compare native implementations with industry-grade solvers, both for mathematical exploration, testing and benchmark purposes.
+
+To give an idea of the interface, the following code allows to integrate the initial value problem
+
+$$
+\dot{x} = f(x,t) = A x, \qquad x(t_0) = x_0
+$$
+
+where $A$ is the 2x2 matrix $(1, -1; 2, -3)$, $x_0 = (-1; 1)$ and $t_0=0$.
+
+```ocaml
+open Owl
+open Owl_ode
+open Owl_ode.Types
+
+(* f(x,t) *)
+let f x t =
+   let a = [|[|1.; -1.|];
+             [|2.; -3.|]|]
+           |> Mat.of_arrays in
+   Mat.(a *@ x)
+
+(* temporal specification:
+   construct a record using the constructor T1 and
[…]
