+++ joss/paper.bib
@@ -1,0 +1,29 @@
+@misc{Wang:2017,
+    title={Owl: A General-Purpose Numerical Library in OCaml},
+    author={Liang Wang},
+    year={2017},
+    eprint={1707.09616},
+    archivePrefix={arXiv},
+    primaryClass={cs.MS},
+    url = {https://arxiv.org/abs/1707.09616}
+}
+
+@incollection{NODE:2018,
+    title = {Neural Ordinary Differential Equations},
+    author = {Chen, Tian Qi and Rubanova, Yulia and Bettencourt, Jesse and Duvenaud, David K},
+    booktitle = {Advances in Neural Information Processing Systems 31},
+    editor = {S. Bengio and H. Wallach and H. Larochelle and K. Grauman and N. Cesa-Bianchi and R. Garnett},
+    pages = {6571--6583},
+    year = {2018},
+    publisher = {Curran Associates, Inc.},
+    url = {http://papers.nips.cc/paper/7892-neural-ordinary-differential-equations.pdf}
+}
+
+@article{VBS:2019,
+    author={Mats Vermeeren and Alessandro Bravetti and Marcello Seri},
+    title={Contact variational integrators},
+    journal={Journal of Physics A: Mathematical and Theoretical},
+    url={http://iopscience.iop.org/10.1088/1751-8121/ab4767},
+    year={2019},
+    doi={10.1088/1751-8121/ab4767}
+}
