### 2019-12-19 #1812 @mseri
Thank you very much for the help. I have added the joss paper to the current release of OwlDE, tagged it and released it. It is now published on Zenodo with DOI 10.5281/zenodo.3584264
