### 2019-12-19 #1812 @mjsottile
:wave: @openjournals/joss-eics This paper has completed the review process and I believe it's ready for final processing.
