### 2019-12-20 #1812 @whedon
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
